// pages/types/documentMetadata.ts
export type Id = string;
export type CreatedAt = string;
export type UpdatedAt = string;
export type DocumentId = string;
export type DocumentType =
  | "BANK_STATEMENT"
  | "PAYSLIP"
  | "UTILITY_BILL"
  | "COUNCIL_TAX"
  | "MORTGAGE_STATEMENT"
  | "CREDIT_CARD_STATEMENT"
  | "INVESTMENT_STATEMENT"
  | "PENSION_STATEMENT"
  | "TAX_DOCUMENT"
  | "OTHER";
export type DocumentMetadata1 = BankStatementData | PayslipData;
// ...
export interface DocumentMetadata {
  id: Id;
  created_at: CreatedAt;
  updated_at: UpdatedAt;
  document_id: DocumentId;
  document_type: DocumentType;
  document_metadata: DocumentMetadata1;
  [k: string]: unknown;
}
export interface BankStatementData {
  iban?: Iban;
  type?: DocumentType1;
  currency: Currency;
  end_balance?: MonetaryAmount | null;
  account_type?: AccountType | null;
  start_balance?: MonetaryAmount | null;
  account_holder?: AccountHolder;
  account_number?: AccountNumber;
  total_money_in?: MonetaryAmount | null;
  bank_identifier: BankIdentifier;
  overdraft_limit?: MonetaryAmount | null;
  total_money_out?: MonetaryAmount | null;
  analysis_results?: AnalysisResults | null;
  statement_period_end?: StatementPeriodEnd;
  statement_period_start?: StatementPeriodStart;
  [k: string]: unknown;
}
// ...
export interface PayslipData {
  type?: DocumentType2;
  net_pay: MonetaryAmount;
  net_ytd?: MonetaryAmount | null;
  currency: Currency;
  gross_pay: MonetaryAmount;
  gross_ytd?: MonetaryAmount | null;
  deductions?: Deductions;
  exemptions?: Exemptions;
  employee_id?: EmployeeId;
  country_code: CountryCode;
  process_date?: ProcessDate;
  employer_name?: EmployerName;
  local_tax_code?: LocalTaxCode;
  pay_period_end?: PayPeriodEnd;
  tax_identifier?: TaxIdentifier;
  employer_tax_id?: EmployerTaxId;
  pay_period_start?: PayPeriodStart;
  employer_registration?: EmployerRegistration;
  [k: string]: unknown;
}